Move rego only test fixtures into the rego tests file
This will avoid linter errors.
